♻️ shim win
