+++ src/data-providers/win.ts
@@ -1,0 +1,3 @@
+const win: { indexedDB?: any; IDBKeyRange?: any } = {}
+
+export default win
